Enviado despues de hacer el despliegue en azure con modulo de balances

+++ ControWell/Shared/AforoTK.cs
@@ -11,7 +11,7 @@ namespace ControWell.Shared
         public int Id { get; set; }
         public Tanque? Tanque { get; set; }
         public int TanqueId { get; set; }
-        public float Nivel { get; set; }
+        public decimal Nivel { get; set; }
         public decimal Volunen { get; set; }
         public int TempBase { get; set; }
         public string Material { get; set; } = string.Empty;
